% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteroidDoseR-package.R
\docType{package}
\name{SteroidDoseR-package}
\alias{SteroidDoseR}
\alias{SteroidDoseR-package}
\title{SteroidDoseR: Corticosteroid Daily Dose Calculation from OMOP CDM Data}
\description{
Provides Baseline (structured OMOP fields) and Rule-based NLP (free-text SIG parsing) methods for computing daily prednisone-equivalent doses from OMOP CDM drug_exposure records. Designed for myositis and other rheumatologic cohort studies. Includes episode building and evaluation against a gold standard.
}
\author{
\strong{Maintainer}: Minqi Xiong \email{mxiong5@jhu.edu}

}
\keyword{internal}
